mock: Refactor
