+++ cmd/manager/main_test.go
@@ -7,9 +7,9 @@ import (
 	"github.com/golang/mock/gomock"
 
 	api "github.com/kubeflow/katib/api"
-	mockdb "github.com/kubeflow/katib/mock/db"
-	mockmodelstore "github.com/kubeflow/katib/mock/modelstore"
-	mockworker "github.com/kubeflow/katib/mock/worker"
+	mockdb "github.com/kubeflow/katib/pkg/mock/db"
+	mockmodelstore "github.com/kubeflow/katib/pkg/mock/modelstore"
+	mockworker "github.com/kubeflow/katib/pkg/mock/worker"
 )
 
 func TestCreateStudy(t *testing.T) {
